Remove old undocumented/untested extension API
This was added way back in #92 as a riff on #64. I like the idea, but
since it's untested and virtually undocumented it has faded into
obscurity even for ensime-vim core developers, I believe. It has never
really been exposed in the modern incarnation of our Neovim rplugin.

We're doing a lot of major refactoring and any dead code like this only
gets in the way. I have some new ideas for a plugin API that could be a
spiritual successor to this, and I'd like to bring such an API back with
documentation, tests, and ideally dogfooding that proves the design and
keeps it working.