Make the polling thread restartable
Killing the server doesn't completely lead to everything restarting yet,
but it's almost there. Might leave the rest to @ktonga on #350.

Generally it's supposed to be an invariant that if client's
self.connected is true, then self.ws should be set--so use that
everywhere instead of having a mix that's confusing.