Keep classpath as a list to support cross platform paths sep
By keeping the classpath as a list, and only generating the
classpath argument at the arugment stage, alternate platform path
separators can be used.  Corresponding changes were
made in the tests to support the change as well.

Additionally, if on a windows platform, a ".exe" extension was added
to the java executable path, if on a windows platform.  This is necessary
to test the existence and execution status as well as a make the
launcher more robust if the user PATHEXT environment variable did
not include the `.exe` extension.